# Conflicts:
#	src/main/kotlin/com/github/nejer6/plugins/Routing.kt

+++ src/main/kotlin/com/github/nejer6/util/naturalsort/Part.kt
@@ -1,6 +1,6 @@
 package com.github.nejer6.util.naturalsort
 
 sealed class Part {
-    data class Number(val value: Int) : Part()
+    data class Number(val value: Long) : Part()
     data class String(val value: kotlin.String) : Part()
 }
